+++ src/Smaragd/Commands/AsyncViewModelCommand.cs
@@ -2,6 +2,7 @@
 using System.ComponentModel;
 using System.Threading.Tasks;
 using System.Windows.Input;
+using NKristek.Smaragd.Helpers;
 using NKristek.Smaragd.ViewModels;
 
 namespace NKristek.Smaragd.Commands
