+++ src/Smaragd/Helpers/ActionDisposable.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace NKristek.Smaragd
+namespace NKristek.Smaragd.Helpers
 {
     /// <inheritdoc />
     /// <remarks>
