+++ src/Smaragd/ViewModels/Bindable.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using NKristek.Smaragd.Helpers;
+using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Runtime.CompilerServices;
