+++ test/Smaragd.Tests/ViewModels/BindableTests.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using NKristek.Smaragd.ViewModels;
+using NKristek.Smaragd.Helpers;
 using Xunit;
 
 namespace NKristek.Smaragd.Tests.ViewModels
